requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.24.3
pyspark==3.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# ETL & Orchestration
apache-airflow==2.7.3
prefect==2.14.11

# Data Visualization
streamlit==1.28.2
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# AI/ML
langchain==0.1.0
openai==1.3.7
langchain-sqlalchemy==0.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
tqdm==4.66.1

# Development
jupyter==1.0.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Optional: Cloud & Scale
boto3==1.34.0
snowflake-connector-python==3.5.0
